⚡︎ feat: migrate from CRA to Vite
