+++ src/common/ErrorDisplay.jsx
@@ -1,12 +1,11 @@
 // @flow
 
-import React from 'react';
-import type { Node, ComponentType } from 'react';
+import * as React from 'react';
 
 type Props = {
   message: string,
 };
-function ErrorDisplay({ message }: Props): Node {
+function ErrorDisplay({ message }: Props): React.Node {
   return (
     <p className="error-block" style={{ color: 'red' }}>
       {message}
@@ -16,4 +15,4 @@ function ErrorDisplay({ message }: Props): Node {
 
 ErrorDisplay.displayName = 'ErrorDisplay';
 
-export const MemoizedErrorDisplay: ComponentType<Props> = React.memo(ErrorDisplay);
+export const MemoizedErrorDisplay: React.ComponentType<Props> = React.memo(ErrorDisplay);
