+++ vite.config.js
@@ -1,0 +1,21 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+
+export default defineConfig({
+  plugins: [
+    react({
+      babel: {
+        presets: ['@babel/preset-flow'],
+        plugins: ['babel-plugin-syntax-hermes-parser'],
+      },
+    }),
+  ],
+  server: {
+    open: true,
+  },
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: './src/setupTests.js',
+  },
+});
